Wisp Suggestion: Immediate Day
I propose we change Wisp to be an Immediate Day effect instead of an Immediate Night effect. 
a) In OS6C it was kind of weird for Wisp to have an immediate ability that they'd have to use as late as possible into the night
b) It's a buff but I don't really think it'd be an issue. If this is used in a full game it's more likely there's 1 Pyromancer and 1 Wisp, so communication between them is probably pretty difficult anyway

Alternatives: Make it an end night ability