Create Juror
This is a group/investigative role that I think can work well in big games. They're essentially bakers who can investigate every other day but have the risk of disbanding if enough of they die/become undead. I'm not sure if this falls in the group category or the investigative category.

To note: The jury is affected by flute player obstructions as long as at-least one juror is enchanted.
To clarify: Only the role is announced publicly, not the player investigated.