Scared Wolf: Remove mention of Runner
The role description claims that ScW is like Runner. That's not really particularly true. The ScW can escape any attack (not just town attacks) and it's escape is a standard escape, not a special one like Runner (which tells the wolves it was a runner). We also got rid of phrases like this in most other places (like e.g. "The Warlock is the werewolf version of the Fortune Teller)