Update lone wolf (#688)
* Update lone wolf

Fixes #674 

I changed it to basically just be the anti priest.

UA and Solo resulting in nothing is better because it gives the LW a purpose besides attacking - they can check UA/Solo claims basically. By only letting it kill townies, and by removing the potential double wolf death the role becomes quite a bit less volatile. Due to the higher amount of townies as opposed to wolves, the LW probably still would want to attack randomly, but I'm unsure how to stop that from being a thing.

* hi im back

* Update lone wolf

replace obstructed, as thats a special term used for something different

Co-authored-by: jay <70105022+cryptidjayden@users.noreply.github.com>